ex09

+++ EX09ComputerShop/src/component/HardDrive.java
@@ -1,16 +1,14 @@
 package component;
 
-import java.math.BigDecimal;
-
 public class HardDrive extends ComputerComponent {
 
     private String name;
     private HardDriveSocket socket;
     private Integer energy;
-    private BigDecimal performancePoints;
+    private Integer performancePoints;
     private Integer price;
 
-    public HardDrive(String name, HardDriveSocket socket, Integer energy, BigDecimal performancePoints, Integer price) {
+    public HardDrive(String name, HardDriveSocket socket, Integer energy, Integer performancePoints, Integer price) {
         this.name = name;
         this.socket = socket;
         this.energy = energy;
@@ -30,7 +28,7 @@ public class HardDrive extends ComputerComponent {
         return energy;
     }
 
-    public BigDecimal getPerformancePoints() {
+    public Integer getPerformancePoints() {
         return performancePoints;
     }
