+++ EX09ComputerShop/src/component/Memory.java
@@ -1,16 +1,14 @@
 package component;
 
-import java.math.BigDecimal;
-
 public class Memory extends ComputerComponent {
 
     private String name;
     private MemorySocket socket;
     private Integer energy;
-    private BigDecimal performancePoints;
+    private Integer performancePoints;
     private Integer price;
 
-    public Memory(String name, MemorySocket socket, Integer energy, BigDecimal performancePoints, Integer price) {
+    public Memory(String name, MemorySocket socket, Integer energy, Integer performancePoints, Integer price) {
         this.name = name;
         this.socket = socket;
         this.energy = energy;
@@ -30,7 +28,7 @@ public class Memory extends ComputerComponent {
         return energy;
     }
 
-    public BigDecimal getPerformancePoints() {
+    public Integer getPerformancePoints() {
         return performancePoints;
     }
